Refactor tilskuddsperioder tabeller: Introduced TilskuddsperiodeRad component and updated InfoRundtTilskuddsperioder for improved structure and readability
