+++ src/AvtaleSide/steg/BeregningTilskudd/visningTilskuddsperioder/VisningTilskuddsperioderTabellVtao.tsx
@@ -1,71 +1,21 @@
 import React, { useContext, useState } from 'react';
-import { BodyShort, Table } from '@navikt/ds-react';
+import { Table } from '@navikt/ds-react';
 import { TilskuddsPeriode } from '@/types/avtale';
-import { formaterDato, formaterPeriode } from '@/utils/datoUtils';
-import EtikettStatus from '@/BeslutterSide/EtikettStatus';
-import { formaterPenger } from '@/utils/PengeUtils';
 import InfoRundtTilskuddsperioder from '@/AvtaleSide/steg/BeregningTilskudd/visningTilskuddsperioder/InfoRundtTilskuddsperioder';
 import BEMHelper from '@/utils/bem';
 import { InnloggetBrukerContext } from '@/InnloggingBoundary/InnloggingBoundary';
 import { AvtaleContext } from '@/AvtaleProvider';
-import { getYear, addDays, isWithinInterval } from 'date-fns';
+import { addDays } from 'date-fns';
 import {
     antallAktiveTilskuddsperioder,
     getIndexVisningForTilskuddsperiode,
 } from '@/AvtaleSide/steg/BeregningTilskudd/visningTilskuddsperioder/visningTilskuddsperiodeUtils';
+import VerticalSpacer from '@/komponenter/layout/VerticalSpacer';
+import TilskuddsperiodeRad from '@/AvtaleSide/steg/BeregningTilskudd/visningTilskuddsperioder/TilskuddsperiodeTabellRad';
 
 interface Properties {
     className: string;
 }
-
-const tilskuddsperiodeRad = ({
-    key,
-    avtaleOpprettet,
-    erNavAnsatt,
-    periode,
-    kreverOppfølgingDato,
-}: {
-    key?: number;
-    avtaleOpprettet: Date;
-    erNavAnsatt: boolean;
-    periode: TilskuddsPeriode;
-    kreverOppfølgingDato?: Date;
-}) => {
-    const periodeAar = getYear(new Date(periode.startDato));
-    // Hvis tilskuddsperioden gjelder for et tidligere år enn når avtalen er opprettet,
-    // så vil vi vise en liten notis om at VTAO-satsen er basert på et lavere beløp
-    const erITidligereAar = periodeAar < getYear(avtaleOpprettet);
-
-    const periodeStatus =
-        kreverOppfølgingDato &&
-        isWithinInterval(kreverOppfølgingDato, { start: periode.startDato, end: periode.sluttDato })
-            ? 'OPPFØLGING_KREVES'
-            : periode.status;
-
-    return (
-        <Table.Row key={key}>
-            <Table.DataCell textSize={'small'}>
-                <BodyShort size="small">{formaterPeriode(periode.startDato, periode.sluttDato)}</BodyShort>
-                {erITidligereAar && (
-                    <BodyShort size="small" textColor="subtle">
-                        Sats for {periodeAar}
-                    </BodyShort>
-                )}
-            </Table.DataCell>
-            {erNavAnsatt && (
-                <Table.DataCell textSize={'small'}>
-                    <EtikettStatus tilskuddsperiodestatus={periodeStatus} size="small" />
-                </Table.DataCell>
-            )}
-            <Table.DataCell align={'right'} textSize={'small'}>
-                {periode.beløp !== null ? formaterPenger(periode.beløp) : '—'}
-            </Table.DataCell>
-            <Table.DataCell textSize={'small'}>
-                {formaterDato(addDays(new Date(periode.sluttDato), 3).toString(), 'dd MMM yyyy')}
-            </Table.DataCell>
-        </Table.Row>
-    );
-};
 
 const VisningTilskuddsperioderTabellVtao: React.FC<Properties> = ({ className }: Properties) => {
     const [visAllePerioder, setVisAllePerioder] = useState(false);
@@ -100,32 +50,33 @@ const VisningTilskuddsperioderTabellVtao: React.FC<Properties> = ({ className }:
                     {avtale.tilskuddPeriode
                         .filter((p: TilskuddsPeriode) => p.aktiv)
                         .slice(startIndexVisning, sluttIndexVisning)
-                        .map((periode: TilskuddsPeriode, key: number) => {
-                            return tilskuddsperiodeRad({
-                                key,
-                                periode,
-                                avtaleOpprettet,
-                                erNavAnsatt,
-                                kreverOppfølgingDato,
-                            });
-                        })}
+                        .map((periode: TilskuddsPeriode, key: number) => (
+                            <TilskuddsperiodeRad
+                                avtaleOpprettet={avtaleOpprettet}
+                                erNavAnsatt={erNavAnsatt}
+                                periode={periode}
+                                key={periode.id}
+                                kreverOppfølgingDato={kreverOppfølgingDato}
+                            />
+                        ))}
                     {!visAllePerioder && sistePeriode && (
                         <>
-                            <Table.Row key={1}>
+                            <Table.Row>
                                 <Table.DataCell textSize={'small'} colSpan={100}>
                                     ...
                                 </Table.DataCell>
                             </Table.Row>
-                            {tilskuddsperiodeRad({
-                                erNavAnsatt,
-                                avtaleOpprettet,
-                                periode: sistePeriode,
-                                kreverOppfølgingDato,
-                            })}
+                            <TilskuddsperiodeRad
+                                avtaleOpprettet={avtaleOpprettet}
+                                periode={sistePeriode}
+                                erNavAnsatt={erNavAnsatt}
+                                kreverOppfølgingDato={kreverOppfølgingDato}
+                            />
                         </>
                     )}
                 </Table.Body>
             </Table>
+            <VerticalSpacer rem={1} />
             <InfoRundtTilskuddsperioder
                 className={cls.className}
                 gjeldendeInnholdStartdato={avtale.gjeldendeInnhold.startDato}
